user-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
redis==5.0.1
alembic==1.12.1
jinja2==3.1.3
aiofiles==23.2.1
python-dotenv==1.0.0
starlette==0.27.0
itsdangerous==2.1.2
httpx==0.26.0
pytest==7.4.4  # 用於測試
requests==2.31.0  # 用於測試
